#3 string id 값 변경 LeagueItemSpaceDecoration 클래스 네이밍 오타 수정 item_league right margin 제거 및 LeagueItemSpaceDecoration에서 item space 추가 MathUtils#getWinRate() 리턴 값 리펙토링

+++ app/src/main/java/space/chuumong/lolstats/utils/MathUtils.kt
@@ -7,7 +7,7 @@ object MathUtils {
             return 0
         }
 
-        return (win.toFloat() / (win + loss).toFloat() * 100).toInt()
+        return (win / (win + loss).toFloat() * 100).toInt()
     }
 }
